requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
PyMySQL==1.0.3
pytesseract==0.3.10
Werkzeug==2.3.3
